Fixed bomb-entity explosion collision & stop enemies from moving during death anim
Instead of checking tile positions for bomb-entity collisions, if we create a bounding box at the explosion's tile position sized to one tile (essentially getting the tile that explosion is happening on's bounding box) and compare that to the entity, we get more accurate bomb-entity explosion collisions.

Now that I'm writing this I'm wondering if we could directly reference the tile's bounding box instead of creating a new one at the bomb's position...
At the very least this works as intended and is more accurate than it's previous version.